requirements.txt
groq
flask
flask-cors
python-dotenv
wikipedia
